Implementing Tree Data Structure
In hindsight, using an array to represent all sets and having arrays of
value occurrences really isn't a good way to go about things. Either I'd
have to create some other data structure (probably a tree anyways) to
keep track of common occurrences (supersets) or just find the common
elements each time I want to eliminate sets matching a particular
subset, which will add so much time.

So yeah, the set representations will all be in a tree now. I figure
it'll be simple enough to index and deal with, and I could learn a lot
from implementing it this way. It should help the program to run quickly
and stuff as well.