Beginning Work on Tree Query Function
I'm sticking to this idea of 'wrapper' functions for the main program
(or some other program) to deal with actual proper elements and stuff so
that the functions that actually do things can be written recursively in
an elegant way. If I had to deal with proper element values, my code
would go crazy from having to deal with keeping track of the current
value within the tree. Instead, we're using 'relative values', so that
the function will see any node in the tree just like any other.

I guess a side effect of this is that each node will have to provide a
count for however many children it has. That's fine.