Recursive Mark Function Keeps Count of Child Nodes
Y'know how I said that in order for this 'relative value' thing to work
the nodes would have to store the number of children they had? That was
just completely inaccurate. The other recursive functions keep track of
the number of children themselves, because you can just do that. So why
shouldn't this one do the same? I like how clean these functions are.