Example Tree, Ignore Already Marked Nodes
If a node is marked, that automatically means that neither it nor its
children matter, so the recursive mark function should not deal with
them.

It's late. I don't feel like writing any more documentation than a
table. I'll do it later. Maybe in the morning I'll actually eat
breakfast.