Polishing up Documentation
I removed some bits in the function description that were overly
in-depth about what operations the mark function is doing. I figured
it'd be better to save it for the code itself, as it's just supposed to
be an overview. I also finished up the main description and updated some
other bits throughout.