Beginning Print Function
I've begun work on the print function. It has a wrapper, a recursive
helper, and another helper for printing out the set itself. The helper
functions use the same relative values as the mark function. Lacking in
documentation right now, but I'll add it later.

Also I changed a couple of types for consistency.