Equivalent Sets Program
Pretty much just copied over some stuff I already had in the main
program before I decided to break everything up. I still need to copy
over and write some documentation for the program as a whole. The only
other thing that needs to be written here is the function for expanding
a set, as well as its helper functions, like for inserting numbers in
order.

So, I've come to the realization that this method of enumerating
equivalent pairs for values 1-M and substituting them in for values
across all known nullifiable sets won't necessarily cover all
nullifiable sets. This is because sometimes when nullifying a set we
might need to go through a value outside the 1-M range. Theoretically,
pretty much all sets are covered though, as it seems like there should
be very few cases where this is the only way to nullify a set. But
ultimately, this is not an exhaustive test on its own, and a separate
function for testing any remaining sets after this program finishes up
will be required.

I have an idea for a better method involving making every possible
arithmetic operation and then letting the results be the set of valid
values for another generation of pairs. This should cover all possible
computations along every step of nullifying a set. However, this just
seems a great deal more expensive and I'd just like to see how effective
the 1-M pairs method is on its own. Perhaps with larger sets it grows
less and less effective and it might be worthwhile to implement
something like that. I dunno. I'll figure it out.