Equivalent Set Enumeration Function
I think that should be the function I want. That was easier to implement
than I thought it'd be. It takes in a set and will enumerate equivalent
sets by substituting in every equivalent pair for every value. I need to
test this to see if it actually works the way I want it to.

Also, I changed some of the types to make everything consistent.