New Haskell Program for Innullifiability
So I'm thinking the Innullifiability program I've been working on might
be using the completely wrong method. It works by eliminating all the
nullifiable sets and then seeing what's left. Sure, the method for
getting to all those nullifiable sets is pretty dang efficient, but I
believe that maybe it's faster to just never bother with the nullifiable
sets at all.

When Vedis and I found the solutions for N = 4, M = 9 in our heads, we
didn't have to do that. What we did was just start from the lowest
number and recursively check what higher numbers could work.

This program will be centred around a single recursive function.
Essentially, it will take in a known innullifiable set, and a list of
all values which can be computed by values inside it. Any numbers in the
range but not in this list can be appended to the set to create a new
innullifiable set. Once we have a set reach the desired set size, it's
put into a list and returned.

I'm not sure if it'll be faster. But I do think that writing it in
Haskell will be the best way in this instance, as we do not need to
store anything. Hopefully this will further expand the range for which
this problem can be solved.